Fix missing change to input group
